more wip on tests
