+++ tests/endpoints.spec.ts
@@ -1,5 +1,27 @@
 import { testAuth } from "./fixtures/auth.fixture";
-import { expect } from "@playwright/test";
+import { test, expect } from "@playwright/test";
+
+test("Unauthorized access to restricted API endpoints using GET", async ({
+	request,
+}) => {
+	const res1 = await request.get("/api/images");
+	expect(res1.status() === 403);
+	const res2 = await request.get("/api/firestore");
+	expect(res2.status() === 403);
+});
+
+test("Unauthorized access to restricted API endpoints using POST", async ({
+	request,
+}) => {
+	const dummy = {
+		title: "dummy",
+		body: "data",
+	};
+	const res1 = await request.post("/api/images", { data: dummy });
+	expect(res1.status() === 403);
+	const res2 = await request.post("/api/firestore", { data: dummy });
+	expect(res2.status() === 403);
+});
 
 testAuth(
 	"Authorized access restricted API endpoints using GET",
